♻️ :: (#33) refactor package name
