+++ application-domain/src/main/java/io/github/v1serviceapplication/extension/Extension.java
@@ -1,0 +1,14 @@
+package io.github.v1serviceapplication.extension;
+
+import io.github.v1serviceapplication.annotation.Aggregate;
+import lombok.*;
+
+import java.util.UUID;
+
+@Getter
+@Builder
+@Aggregate
+public class Extension {
+    private UUID userId;
+    private UUID studyRoomId;
+}
