--- application-domain/src/main/java/io/github/v1serviceapplication/extensionapply/Extension.java
@@ -1,14 +1,0 @@
-package io.github.v1serviceapplication.extensionapply;
-
-import io.github.v1serviceapplication.annotation.Aggregate;
-import lombok.*;
-
-import java.util.UUID;
-
-@Getter
-@Builder
-@Aggregate
-public class Extension {
-    private UUID userId;
-    private UUID studyRoomId;
-}
