--- application-domain/src/main/java/io/github/v1serviceapplication/picnicapply/Picnic.java
@@ -1,20 +1,0 @@
-package io.github.v1serviceapplication.picnicapply;
-
-import io.github.v1serviceapplication.annotation.Aggregate;
-import lombok.*;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.UUID;
-
-@Aggregate
-@Getter
-@Builder
-public class Picnic {
-    private UUID userId;
-    private LocalDate date;
-    private LocalTime startTime;
-    private LocalTime endTime;
-    private String reason;
-    private String arrangement;
-}
