+++ application-domain/src/main/java/io/github/v1serviceapplication/stay/Stay.java
@@ -1,0 +1,14 @@
+package io.github.v1serviceapplication.stay;
+
+import io.github.v1serviceapplication.annotation.Aggregate;
+import lombok.*;
+
+import java.util.UUID;
+
+@Aggregate
+@Getter
+@Builder
+public class Stay {
+    private UUID userId;
+    private String code;
+}
