--- application-domain/src/main/java/io/github/v1serviceapplication/stayapply/Stay.java
@@ -1,14 +1,0 @@
-package io.github.v1serviceapplication.stayapply;
-
-import io.github.v1serviceapplication.annotation.Aggregate;
-import lombok.*;
-
-import java.util.UUID;
-
-@Aggregate
-@Getter
-@Builder
-public class Stay {
-    private UUID userId;
-    private String code;
-}
