+++ application-domain/src/main/java/io/github/v1serviceapplication/weekendmeal/WeekendMeal.java
@@ -1,0 +1,14 @@
+package io.github.v1serviceapplication.weekendmeal;
+
+import io.github.v1serviceapplication.annotation.Aggregate;
+import lombok.*;
+
+import java.util.UUID;
+
+@Aggregate
+@Getter
+@Builder
+public class WeekendMeal {
+    private UUID userId;
+    private Boolean isApplied;
+}
