--- application-domain/src/main/java/io/github/v1serviceapplication/weekendmealapply/WeekendMeal.java
@@ -1,14 +1,0 @@
-package io.github.v1serviceapplication.weekendmealapply;
-
-import io.github.v1serviceapplication.annotation.Aggregate;
-import lombok.*;
-
-import java.util.UUID;
-
-@Aggregate
-@Getter
-@Builder
-public class WeekendMeal {
-    private UUID userId;
-    private Boolean isApplied;
-}
