--- application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/extensionapply/domain/repository/ExtensionRepository.java
@@ -1,11 +1,0 @@
-package io.github.v1serviceapplication.domain.extensionapply.domain.repository;
-
-import io.github.v1serviceapplication.domain.extensionapply.domain.ExtensionEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface ExtensionRepository extends JpaRepository<ExtensionEntity, UUID> {
-}
