+++ application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/picnic/domain/PicnicEntity.java
@@ -1,0 +1,31 @@
+package io.github.v1serviceapplication.domain.picnic.domain;
+
+import io.github.v1serviceapplication.global.entity.BaseUUIDEntity;
+import lombok.*;
+
+import javax.persistence.*;
+import javax.validation.constraints.NotNull;
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Builder
+@Table(name = "tbl_picnic_apply")
+@Entity
+public class PicnicEntity extends BaseUUIDEntity {
+    @NotNull
+    private LocalDate date;
+
+    @NotNull
+    private LocalTime startTime;
+
+    private LocalTime endTime;
+
+    @NotNull
+    private String reason;
+
+    @NotNull
+    private String arrangement;
+}
