--- application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/picnicapply/domain/repository/PicnicRepository.java
@@ -1,11 +1,0 @@
-package io.github.v1serviceapplication.domain.picnicapply.domain.repository;
-
-import io.github.v1serviceapplication.domain.picnicapply.domain.PicnicEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface PicnicRepository extends JpaRepository<PicnicEntity, UUID> {
-}
