+++ application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/stay/StayConfig.java
@@ -1,0 +1,16 @@
+package io.github.v1serviceapplication.domain.stay;
+
+import io.github.v1serviceapplication.annotation.DomainService;
+import io.github.v1serviceapplication.stay.Stay;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.FilterType;
+
+@Configuration
+@ComponentScan(
+        basePackageClasses = {Stay.class},
+        includeFilters = { @ComponentScan.Filter(
+        type = FilterType.ANNOTATION,
+        value = DomainService.class
+)})
+public class StayConfig {}
