+++ application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/stay/domain/repository/StayRepository.java
@@ -1,0 +1,11 @@
+package io.github.v1serviceapplication.domain.stay.domain.repository;
+
+import io.github.v1serviceapplication.domain.stay.domain.StayEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface StayRepository extends JpaRepository<StayEntity, UUID> {
+}
