+++ application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/stay/mapper/StayMapper.java
@@ -1,0 +1,12 @@
+package io.github.v1serviceapplication.domain.stay.mapper;
+
+import io.github.v1serviceapplication.domain.stay.domain.StayEntity;
+import io.github.v1serviceapplication.stay.Stay;
+import org.mapstruct.Mapper;
+
+@Mapper
+public interface StayMapper {
+    StayEntity stayDomainToEntity(Stay stay);
+
+    Stay stayEntityToDomain(StayEntity stayEntity);
+}
