+++ application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/stay/presentation/StayController.java
@@ -1,0 +1,11 @@
+package io.github.v1serviceapplication.domain.stay.presentation;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/stay")
+public class StayController {
+}
