--- application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/stayapply/domain/repository/StayRepository.java
@@ -1,11 +1,0 @@
-package io.github.v1serviceapplication.domain.stayapply.domain.repository;
-
-import io.github.v1serviceapplication.domain.stayapply.domain.StayEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface StayRepository extends JpaRepository<StayEntity, UUID> {
-}
