--- application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/stayapply/presentation/StayController.java
@@ -1,11 +1,0 @@
-package io.github.v1serviceapplication.domain.stayapply.presentation;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/stay")
-public class StayController {
-}
