+++ application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/weekendmeal/domain/repository/WeekendMealRepository.java
@@ -1,0 +1,11 @@
+package io.github.v1serviceapplication.domain.weekendmeal.domain.repository;
+
+import io.github.v1serviceapplication.domain.weekendmeal.domain.WeekendMealEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface WeekendMealRepository extends JpaRepository<WeekendMealEntity, UUID> {
+}
